refactor(log): move masking logic from grpc.go to log.go for better organization
- Relocate maskBinaryFields and its helpers from grpc.go to log.go.
- Consolidate logging-related utilities into a more appropriate file.
- No functional changes, purely moving code for better maintainability.
